+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [v0.3.0] - 2022-04-29
+
 ### Changed
 
 - [breaking-change] Made `Hasher` a subtrait of `core::hash::Hasher`, and
@@ -48,7 +50,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Initial release
 
-[Unreleased]: https://github.com/japaric/hash32/compare/v0.2.1...HEAD
+[Unreleased]: https://github.com/japaric/hash32/compare/v0.3.0...HEAD
+[v0.3.0]: https://github.com/japaric/hash32/compare/v0.2.1...v0.3.0
 [v0.2.1]: https://github.com/japaric/hash32/compare/v0.2.0...v0.2.1
 [v0.2.0]: https://github.com/japaric/hash32/compare/v0.1.2...v0.2.0
 [v0.1.2]: https://github.com/japaric/hash32/compare/v0.1.1...v0.1.2